% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ab.est}
\alias{ab.est}
\title{Estimate the parameters via ERCC spike-in for TASC model.}
\usage{
ab.est(x, y)
}
\arguments{
\item{x}{A vector indicating molecule numbers of ERCC spike-in's in each cell.}

\item{y}{A matrix or data frame, with each row indicating the read counts of ERCC spike-in's in each cell.}
}
\value{
A two-column data frame including the estimated Alpha and Beta for each cell.
}
\description{
The parameters Alpha and Beta for each cell is estimated by linear regression.
}
\author{
Rong Ma, University of Pennsylvania <rongm@mail.med.upenn.edu>
}

