% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{icet.plot}
\alias{icet.plot}
\title{Cell Type Visualization via IceT}
\usage{
icet.plot(icet, pch = 20, reset = FALSE, recover.data = NA,
  imputation = FALSE, max_iter = 1000, perplexity = 50, eps = 2,
  MinPts = 5, set.color = NA)
}
\arguments{
\item{icet}{The list object returned from the icet function.}

\item{pch}{A graphical parameter, specifying the characteristics of plotting points.}

\item{reset}{A logical (default: FALSE), if TRUE, the user can set other parameters provided the recover.data is included.}

\item{recover.data}{The recovered data generated by recover function; or imputated data. Used only when reset = TRUE.}

\item{imputation}{A logical (default: FALSE) indicating whether the imputated data is used. Used only when reset = TRUE.}

\item{max_iter}{integer; Number of iterations in Rtsne function (default: 1000). Can only be changed when reset = TRUE.}

\item{eps}{Reachability distance in dbscan function (default: 2). Can only be changed when reset = TRUE.}

\item{MinPts}{Reachability minimum no. of points in dbscan function (default: 5). Can only be changed when reset = TRUE.}

\item{set.color}{Prescribe the colors for the data representations. Defaul as the colors by IceT.}

\item{numeric;}{Perplexity parameter in Rtsne function (default: 50). Can only be changed when reset = TRUE.}
}
\description{
Make plots based on results of IceT.
}
\author{
Rong Ma, University of Pennsylvania <rongm@mail.med.upenn.edu>
}

