% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{icet}
\alias{icet}
\title{Identification of Cell Types using nonparametric density estimation}
\usage{
icet(recover.data, ratio = 20, threshold = 0.5, fdr = 0.002,
  imputation = FALSE, pca = TRUE, max_iter = 1000, perplexity = 50,
  eps = 2, MinPts = 5)
}
\arguments{
\item{recover.data}{The recovered or imputated gene expression data set (can be data.frame or matrix) with columns indexing the cell id and the rows indexing the genes.}

\item{ratio}{Numeric; the signal-to-noise ratio thresholding the noise level.}

\item{threshold}{Numeric between 0 to 1; the threshold that eliminate all the genes below that quantile in terms of mean expression levels (default: 0.5)}

\item{fdr}{Numeric; the false discovery rate used in bandwith selection, the smaller fdr is the less genes would be selected in the end (default: 0.002)}

\item{imputation}{A logical indicating whether the imputated data is used (default: FALSE). If `FALSE`, the list object generated from the `recover` function in IceT package should be used.}

\item{pca}{logical; Whether an initial PCA step should be performed (default: TRUE)}

\item{max_iter}{integer; Number of iterations in Rtsne function (default: 1000)}

\item{eps}{Reachability distance in dbscan function (default: 2)}

\item{MinPts}{Reachability minimum no. of points in dbscan function (default: 5)}

\item{numeric;}{Perplexity parameter in Rtsne function (default: 50)}
}
\value{
points A matrix indicating the coordinates of the visualization of the cells on a 2 dimensional plane.

cluster A vector indicating the group index for each cell.

selected_genes A vector of the index of the informative genes selected by IceT.

num_of_clusters A numeric value indicating the number of clusters generated by IceT (including singletons as an extra cluster).
}
\description{
Nonparametric density estimation is used for cell type clustering. The intermediate parameters such as bandwith and signal-to-noise ratio are chosen by build-in algorithms but can also be user-determined.
}
\author{
Rong Ma, University of Pennsylvania <rongm@mail.med.upenn.edu>
}

