% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recover}
\alias{recover}
\title{Recover the gene expression level using ERCC spike-in.}
\usage{
recover(X, est.ab)
}
\arguments{
\item{X}{The gene expression data set (can be data.frame or matrix) with columns indexing the cell id and the rows indexing the genes.}

\item{est.ab}{A two-column data frame including the estimated Alpha and Beta for each cell.}
}
\value{
A list including the recovered gene expression data matrix and its logarithmic transformed version, the reduced gene expression data (with 0's substituted by NA's) and its logarithmic transformed version, with the same structure as the input data set X.
}
\description{
Using the TASC model introduced by Jia and Li (2016), with parameters Alpha and Beta for each cell estimated from ERCC spike-in, we can recover the non-zero counts while ignoring the zero counts. The preprocessed data is readily be used for downstream cell type clustering.
}
\author{
Rong Ma, University of Pennsylvania <rongm@mail.med.upenn.edu>
}

